Added support for pushing a step forward (to be after the next step), and for kicking a step up (to be after its parent).
Added 'cleanup' field to Task, so that it prompts when a step with that task is finished to ensure everything is put away.

Fixes to adding required tasks around other tasks with frozen positions.

If at the start or end of a task, new required tasks are set to have their position frozen by default.

We now log when a task is last completed.

Better fetching of data on new session.

Improvements/minor overhaul of nested task tabs; now hierarchy is maintained even when selected outside of the task tree.

Removed artifacts from cached filters

Made base View class 'div' to allow for easier inheritance later on.